[GH-90] move decorators into single namespace: wrapper
