+++ impyute/util/__init__.py
@@ -2,8 +2,8 @@
 
 from .find_null import find_null
 from .errors import BadInputError
-from .checks import checks
-from .preprocess import preprocess
+from .wrapper import checks
+from .wrapper import preprocess
 from . import inverse_distance_weighting
 
 __all__ = ["find_null", "checks", "BadInputError", "preprocess",
